requirements.txt
babel==2.17.0
bcrypt==4.3.0
blinker==1.9.0
capstone==6.0.0a4
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
colored-traceback==0.4.2
cryptography==45.0.5
Flask==2.3.2
flask-babel==4.0.0
gevent==25.5.1
git-filter-repo==2.47.0
greenlet==3.2.3
idna==3.10
intervaltree==3.1.0
IPy==1.1
itsdangerous==2.2.0
Jinja2==3.1.2
loguru==0.7.3
lxml==6.0.0
Mako==1.3.10
MarkupSafe==3.0.2
ndjson==0.3.1
gunicorn==20.1.0
packaging==25.0
paramiko==3.5.1
plumbum==1.9.0
portscan==1.1
psutil==7.0.0
pwn==1.0
pwntools==4.14.1
pycparser==2.22
pycryptodome==3.20.0
pyelftools==0.32
Pygments==2.19.2
PyNaCl==1.5.0
pyOpenSSL==25.1.0
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-nmap==0.7.1
pytz==2025.2
requests==2.31.0
ROPGadget==7.6
rpyc==6.0.2
setuptools==80.9.0
six==1.17.0
sortedcontainers==2.4.0
tzdata==2025.2
tzlocal==5.3.1
unicorn==2.1.2
unix-ar==0.2.1
urllib3==2.2.1
Werkzeug==2.3.6
win32_setctime==1.2.0
xmltodict==0.13.0
zope.event==5.1
zope.interface==7.2
zstandard==0.23.0
user-agents==2.2.0
python-telegram-bot==20.8
flask-limiter>=3.0
redis>=4.0

